Crate shuffled_iter

This crate provides methods to iterate over a group of values in random order, without allocation and shuffling them all.

Such an iterator may be obtained via the ShuffledIterGen trait.

Structs

ShuffledIter
The actual iterator which iterates over, in random order, all u32s smaller than or equal to a given maximum value.

Traits

ShuffledIterGen
This trait may be used to obtain an instance of ShuffledIter. It should always be implemented generically for all types implementing rand::Rng, that is for all random number generators (which is the case for this library).