Trait shuffled_iter::ShuffledIterGen

pub trait ShuffledIterGen<T>: Rng {
    type Iter: Iterator;
    fn iter_shuffled(&mut self, T) -> Self::Iter;
}

This trait may be used to obtain an instance of ShuffledIter. It should always be implemented generically for all types implementing rand::Rng, that is for all random number generators (which is the case for this library).

As such, an instance of ShuffledIter may be obtained by calling iter_shuffled with a suitable argument on any random number generator, given that ShuffledIterGen is in scope.

An implementation of this trait is provided by this library for the following agument types:

all ranges of integer types (ie. 0i8.., ..10u16, 10..20)

slices of any kind

Associated Types

type Iter: Iterator

Required Methods

fn iter_shuffled(&mut self, T) -> Self::Iter